# docs/swagger.yaml
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  models.Category:
    properties:
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      sold_product_amount:
        type: integer
      type:
        type: string
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      price:
        type: integer
      stock:
        type: integer
      title:
        type: string
      transaction_history:
        items:
          $ref: '#/definitions/models.TransactionHistory'
        type: array
    type: object
  models.TransactionHistory:
    properties:
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: integer
      quantity:
        type: integer
      total_price:
        type: integer
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      balance:
        type: integer
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      role:
        type: string
      transaction_history:
        items:
          $ref: '#/definitions/models.TransactionHistory'
        type: array
    type: object
info:
  contact: {}
paths:
  /categories:
    get:
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Gets all categories
    post:
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category type
        in: body
        name: type
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Creates a new category
  /categories/{id}:
    delete:
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Deletes a category by ID
    patch:
      parameters:
      - description: Bearer token for authentication
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Type
        in: body
        name: type
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated category
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Updates a category type by ID
...
  /users/register:
    post:
      parameters:
      - description: Balance
        in: body
        name: email
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.User'
      summary: Updates the user's balance
swagger: "2.0"